Separate SimpleMenu and MegaMenu test-site token files
