+++ examples/test-site/src/components/MegaMenu/MegaMenu.tsx
@@ -23,7 +23,7 @@ import {
 import { asReadOnly } from '@bodiless/core';
 import { asEditable as withEditorSimple } from '@bodiless/components';
 // import { withEditorSimple } from '../Editors';
-import { withMenuToutStyles, withMegaMenuStyles } from './MegaMenu.token';
+import withMegaMenuStyles, { withMenuToutStyles } from './MegaMenu.token';
 import { withToutEditors } from '../Tout';
 
 export const withTitleEditor = withEditorSimple('text', 'Menu Item');
